Minor change to handle unsigned arity more cleanly/simply.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8954 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/builtin.c
@@ -922,8 +922,8 @@ int count_variable_occurrences(Cell term) {
   case XSB_ATTV:
     return count + 1;
   case XSB_STRUCT: {
-    int i, arity;
-    arity = (int) get_arity(get_str_psc(term));
+    unsigned int i, arity;
+    arity = get_arity(get_str_psc(term));
     if (arity == 0) {
       return count;
     }
